pyproject.toml
[project]
name = "rl-detector"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "tinker",
    "openai>=1.0",
    "datasets",
    "torch",
    "transformers",
    "python-dotenv",
    "pyyaml",
    "tqdm",
    "wandb",
    "weave>=0.52.36",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rl_detector"]
